Use common test fixture
