+++ rosbag2/test/rosbag2/rosbag2_integration_test.cpp
@@ -29,41 +29,11 @@
 #include "rosbag2/rosbag2.hpp"
 #include "std_msgs/msg/string.hpp"
 
+#include "rosbag2_test_fixture.cpp"
+
 using namespace ::testing;  // NOLINT
 using namespace rosbag2;  // NOLINT
 using namespace std::chrono_literals;  // NOLINT
-
-class Rosbag2TestFixture : public Test
-{
-public:
-  Rosbag2TestFixture()
-  : database_name_(UnitTest::GetInstance()->current_test_info()->name())
-  {}
-
-  ~Rosbag2TestFixture() override
-  {
-    std::remove(database_name_.c_str());
-  }
-
-  std::vector<std::string> getMessages(sqlite3 * db, std::string table = "messages")
-  {
-    std::vector<std::string> table_msgs;
-    sqlite3_stmt * statement;
-    std::string query = "SELECT * FROM " + table;
-    sqlite3_prepare_v2(db, query.c_str(), -1, &statement, nullptr);
-    int result = sqlite3_step(statement);
-    while (result == SQLITE_ROW) {
-      table_msgs.emplace_back(
-        std::string(reinterpret_cast<const char *>(sqlite3_column_text(statement, 1))));
-      result = sqlite3_step(statement);
-    }
-    sqlite3_finalize(statement);
-
-    return table_msgs;
-  }
-
-  std::string database_name_;
-};
 
 void recordMessage(
   std::string db_name, std::promise<void> * recorder_promise, std::promise<void> * rcl_promise)
